pyproject.toml
[project]
name = "overwatcher"
version = "0.1.0"
description = "SMS accountability agent for one person."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlmodel>=0.0.22",
    "apscheduler>=3.10,<4",
    "twilio>=9.3",
    "litellm>=1.52",
    "instructor>=1.6",
    "gspread>=6.1",
    "google-auth>=2.35",
    "python-multipart>=0.0.12",
    "httpx>=0.27",
    "python-json-logger>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/overwatcher"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
